Module quill_type_deduce::hir::expr

Structs

AbstractionVariable

A variable bound by some abstraction e.g. a lambda or a let inside it.

BoundVariable

A variable bound by the definition of a function.

DefinitionCaseGeneric

Represents a case of a definition in an impl expression.

Expression

The Expression type is central to the HIR, or high-level intermediate representation. In an expression in HIR, the type of each object is known.

ExpressionT

Enums

ExpressionContentsGeneric

Represents the contents of an expression (which may or may not have been already type checked). The type V represents the type variables that we are substituting into this symbol. The type P represents a pattern. The type I represents the contents of an impl expression. You should use ExpressionContents or ExpressionContentsT instead of this enum directly.

TypeVariable

Closely tied to the Type struct, this is used while type checking to allow unknown types, represented by TypeVariableIds.

Type Definitions

ExpressionContents

ExpressionContentsT