Enum quill_type_deduce::hir::expr::ExpressionContentsGeneric

pub enum ExpressionContentsGeneric<E, T, V, I> {
    Argument(NameP),
    Local(NameP),
    Symbol {
        name: QualifiedName,
        range: Range,
        type_variables: V,
    },
    Apply(Box<E>, Box<E>),
    Lambda {
        lambda_token: Range,
        params: Vec<(NameP, T)>,
        expr: Box<E>,
    },
    Let {
        let_token: Range,
        name: NameP,
        expr: Box<E>,
    },
    Block {
        open_bracket: Range,
        close_bracket: Range,
        statements: Vec<E>,
    },
    ConstructData {
        data_type_name: QualifiedName,
        variant: Option<String>,
        fields: Vec<(NameP, E)>,
        open_brace: Range,
        close_brace: Range,
    },
    ConstantValue {
        value: ConstantValue,
        range: Range,
    },
    Borrow {
        borrow_token: Range,
        expr: Box<E>,
    },
    Copy {
        copy_token: Range,
        expr: Box<E>,
    },
    Impl {
        impl_token: Range,
        implementations: I,
    },
}

Represents the contents of an expression (which may or may not have been already type checked). The type V represents the type variables that we are substituting into this symbol. The type I represents the contents of an impl expression. You should use ExpressionContents or ExpressionContentsT instead of this enum directly.

Variants

Argument(NameP)

An argument to this function e.g. x.

Local(NameP)

A local variable declared by a lambda or let expression.

Symbol

A symbol in global scope e.g. + or fold.

Fields of Symbol

name: QualifiedName

The name that the symbol refers to.

range: Range

The range where the symbol’s name was written in this file.

type_variables: V

The type variables we’re substituting into this symbol. If using an ExpressionT, this should be a vector of TypeVariable. If using an Expression, this should be a vector of Type.

Apply(Box<E>, Box<E>)

Apply the left hand side to the right hand side, e.g. a b. More complicated expressions e.g. a b c d can be desugared into ((a b) c) d.

Lambda

A lambda abstraction, specifically lambda params -> expr.

Fields of Lambda

lambda_token: Rangeparams: Vec<(NameP, T)>expr: Box<E>

Let

A let statement, specifically let identifier = expr;.

Fields of Let

let_token: Rangename: NamePexpr: Box<E>

Block

A block of statements, inside parentheses, separated by line breaks or commas, where the final expression in the block is the type of the block as a whole.

Fields of Block

open_bracket: Rangeclose_bracket: Rangestatements: Vec<E>

ConstructData

Explicitly create a value of a data type.

Fields of ConstructData

data_type_name: QualifiedNamevariant: Option<String>fields: Vec<(NameP, E)>open_brace: Rangeclose_brace: Range

ConstantValue

A raw value, such as a string literal, the unit keyword, or an integer literal.

Fields of ConstantValue

value: ConstantValuerange: Range

Borrow

A borrowed value.

Fields of Borrow

borrow_token: Rangeexpr: Box<E>

Copy

A copy of a borrowed value.

Fields of Copy

copy_token: Rangeexpr: Box<E>

Impl

An implementation of an aspect.

Fields of Impl

impl_token: Range

Maps names of definitions to their implementations.

implementations: I