Module quill_type_deduce::type_check

Performs type deduction and type checking of expressions and patterns.

Structs

ForeignDeclaration

Represents a declaration that may be in a different source file.

TypeVariablePrinter

A utility for printing type variables to screen. Works like the Display trait, but works better for printing type variable names.

VisibleNames

What names are visible in a given source file? This is generated from the imports provided at the top of a source file.

Functions

check

get_args_of_type

Treating this symbol as a function, what are its arguments’ types and the result type? If this is not a function, then it is treated as a zero-argument function.