Enum quill_type_deduce::hir::expr::ExpressionContentsGeneric

pub enum ExpressionContentsGeneric<E, T, V, P, I> {

    Local(NameP),
    Symbol {
        name: QualifiedName,
        range: Range,
        type_variables: V,
    },
    Apply(Box<E>, Box<E>),
    Lambda {
        lambda_token: Range,
        params: Vec<(NameP, T)>,
        expr: Box<E>,
    },
    Let {
        let_token: Range,
        name: NameP,
        expr: Box<E>,
    },
    Block {
        open_bracket: Range,
        close_bracket: Range,
        statements: Vec<E>,
    },
    ConstructData {
        data_type_name: QualifiedName,
        variant: Option<String>,
        fields: Vec<(NameP, E)>,
        open_brace: Range,
        close_brace: Range,
    },
    ConstantValue {
        value: ConstantValue,
        range: Range,
    },
    String {
        value: String,
        range: Range,
    },
    Borrow {
        borrow_token: Range,
        expr: Box<E>,
    },
    Copy {
        copy_token: Range,
        expr: Box<E>,
    },
    Impl {
        impl_token: Range,
        implementations: I,
    },
    Match {
        match_token: Range,
        expr: Box<E>,
        cases: Vec<(P, E)>,
    },

}

Represents the contents of an expression (which may or may not have been already type checked). The type V represents the type variables that we are substituting into this symbol. The type P represents a pattern. The type I represents the contents of an impl expression. You should use ExpressionContents or ExpressionContentsT instead of this enum directly.

Variants

Local(NameP)

Tuple Fields

0: NameP

An argument to this function, or a local variable declared by a lambda or let expression.

Symbol

Fields

name: QualifiedName

The name that the symbol refers to.

range: Range

The range where the symbol’s name was written in this file.

type_variables: V

The type variables we’re substituting into this symbol. If using an ExpressionT, this should be a vector of TypeVariable. If using an Expression, this should be a vector of Type.

A symbol in global scope e.g. + or fold.

Apply(Box<E>, Box<E>)

Tuple Fields

0: Box<E>

1: Box<E>

Apply the left hand side to the right hand side, e.g. a b. More complicated expressions e.g. a b c d can be desugared into ((a b) c) d.

Lambda

Fields

lambda_token: Range

params: Vec<(NameP, T)>

expr: Box<E>

A lambda abstraction, specifically lambda params -> expr.

Let

Fields

let_token: Range

name: NameP

expr: Box<E>

A let statement, specifically let identifier = expr;.

Block

Fields

open_bracket: Range

close_bracket: Range

statements: Vec<E>

A block of statements, inside parentheses, separated by line breaks or commas, where the final expression in the block is the type of the block as a whole.

ConstructData

Fields

data_type_name: QualifiedName

variant: Option<String>

fields: Vec<(NameP, E)>

open_brace: Range

close_brace: Range

Explicitly create a value of a data type.

ConstantValue

Fields

value: ConstantValue

range: Range

A raw value, such as a character literal, the unit keyword, or an integer literal.

String

Fields

value: String

range: Range

A string literal, which will be converted into a cons-list later.

Borrow

Fields

borrow_token: Range

expr: Box<E>

A borrowed value.

Copy

Fields

copy_token: Range

expr: Box<E>

A copy of a borrowed value.

Impl

Fields

impl_token: Range

implementations: I

Maps names of definitions to their implementations.

An implementation of an aspect.

Match

Fields

match_token: Range

expr: Box<E>

cases: Vec<(P, E)>

A list of patterns and their replacements.

A match expression, specifically something of the form match expr { pat -> result, pat -> result, ... }